requirements.txt
replicate==0.18.1
Pillow==10.1.0
tqdm==4.66.1
python-dotenv==1.0.0
requests